Replace DevServer with BlazorGateway in E2E test infrastructure
- Extract BlazorGateway.BuildWebHost() into BlazorGateway.cs (shared source)
- Update Components.TestServer and BlazorWasmTestAppFixture to use BlazorGateway
- Add YARP packages (Yarp.ReverseProxy, ServiceDiscovery) via repo's <Reference> pattern
- Add SpaFallback.targets to BasicTestApp, ThreadingApp, and Wasm.Performance.TestApp
- Remove Program.BuildWebHost.cs Compile Include from E2ETests (gets it via TestServer ref)
