+++ eng/Dependencies.props
@@ -246,6 +246,8 @@ may be turned into `<PackageReference>` items in projects.
     <LatestPackageReference Include="xunit.v3.extensibility.core" Version="$(XunitV3ExtensibilityCoreVersion)" />
     <LatestPackageReference Include="Microsoft.Playwright.Xunit.v3" Version="$(MicrosoftPlaywrightXunitV3Version)" />
     <LatestPackageReference Include="Yarp.ReverseProxy" Version="$(YarpReverseProxyVersion)" />
+    <LatestPackageReference Include="Microsoft.Extensions.ServiceDiscovery" />
+    <LatestPackageReference Include="Microsoft.Extensions.ServiceDiscovery.Yarp" />
   </ItemGroup>
 
   <ItemGroup Label="Dependencies with versions.">
