+++ src/Components/benchmarkapps/Wasm.Performance/TestApp/Wasm.Performance.TestApp.csproj
@@ -13,4 +13,7 @@
   <ItemGroup>
     <Reference Include="Microsoft.AspNetCore.Components.WebAssembly" />
   </ItemGroup>
+
+  <Import Project="$(RepoRoot)src\Components\WebAssembly\testassets\SpaFallback.targets" />
+
 </Project>
